+++ packages/select/src/index.ts
@@ -1,1 +1,5 @@
-export * from './select';
+export * from './root';
+export * from './trigger';
+export * from './options';
+export * from './option';
+export * from './label';
